Move producer bits to consumer module.
